add (2) files: cdm.bash, cdm_function.bash.in
This commit is intended to capture 'cdm' in the form that I have
been using for a while ("eating my own dogfood", as it were). It
works fine, and I have not yet tripped over any major annoyances or
deficiencies.

Note that the meat of the file 'cdm_function.bash.in' was extracted
from a snippet of my personal ~/.bashrc file, where I have the path
to the 'cdm.bash' file hardcoded. If we ever get some GNU Autotools
'configure'-style machinery for this project, that file will need to
be updated for correct configure-time or built-time filtering. As it
stands, it can mostly be copy/pasted into a Bash config file (e.g.,
~/.bash_profile or ~/.bashrc or whatever) and touched-up lightly by
editing the obvious.

Signed-off-by: Alan D. Salewski <ads@salewski.email>