issue #2: cdm: allow multiple directory (sub)trees to be specified on the command line, and 'cd' into the last
This allows for common command line patterns that previously
required multiple commands, such as:

    $ cdm YYYY/MM/DD/some-project/{pristine,munge}
    [user is now in the directory 'YYYY/MM/DD/some-project/munge/']

Prior to this change, 'cdm' got us only part of the way there. To
achieve the same effect the user would have needed to do this:

    $ cdm YYYY/MM/DD/some-project
    [user is now in the directory 'YYYY/MM/DD/some-project/']

    $ mkdir pristine munge
    [two subdirectories created]

    $ cd munge

Backward compatibility has been maintained (except for the
pathological case in which a caller was depending on 'cdm'
erroring-out when provided with more than one command line argument,
but nobody is doing that).

Since we do not yet have any proper configure/build process, in the
'cdm.bash' file we simply bump the hard-coded version number from
0.0.1 to 0.1.0 (which is a semver-compatible indication of new
feature available, but we are otherwise backward compatible with the
earlier version).

Signed-off-by: Alan D. Salewski <ads@salewski.email>